Trying to mock WebMidi and do a test of the 'whole' app.
Something hangs when <App> is initialised in a test - the test never
completes.  Not sure if it's due to useEffect somewhere.
There is a similar test with a separate <Wrapper> component just to
compare, but this seems to run / start okay.
The original idea / goal was to run the whole app and verify what was
being sent to WebMidi